model/requirements.txt
jupyter
numpy
imageio
torch
torchvision
IPython
scikit-image
scipy
matplotlib
opencv-python
streamlit
flask
Pillow